Move resolve-css-selector.js to webview folder
